Fix a critical race condition in XRL method lookup.
Introduce the use of Boost's weak_ptr<T>/shared_ptr<T> for managing
the lifecycle of XrlPFSender objects, which represent open XRL transports.

class Xrl now holds a boost::weak_ptr<XrlPFSender> for its cached, resolved
sender object. This fixes the issue where there was no way of
cleanly invalidating an existing XrlPFSender* transport pointer.

The method in XrlRouter which tries to find an existing RPC endpoint for
outgoing XRLs, has been renamed lookup_sender(), so as not to clash with
the hooks for batch XRL support (which are not currently used).

As most of the XORP code base is not yet thread or Boost aware, we
continue to use XrlPFSender* in most situations.

This change is cumulative with the auto_ptr<T> fix used in the XIF
client-side RPC stubs themselves; this is necessary to ensure the
weak_ptr<T> is private to each XIF client stub instantiation.

This change passes 'scons check' on FreeBSD 8-STABLE/amd64.

Reported by:	Ben Greear, Li Zhao


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11694 26435e26-4b25-483a-942d-477236992e84