license:  Submitters explicitly own copyright on submitted code.
I propose to update the LICENSE file to explicitly note that contributions
on or after today are copyright by the submitter unless otherwise specified
in the commit message and/or committed code.

This ensures that no single entity can re-license XORP or otherwise
take changes in the public svn tree and use or relicense them in ways not compatible
with the current license scheme (GPL, LGPL, other) without consent
of all committers from today on.  I believe this will help make
xorp stronger and it will certainly make me happier about committing
my changes to the public svn tree.

Signed-off-by: Ben Greear <greearb@candelatech.com>

git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@11697 26435e26-4b25-483a-942d-477236992e84