policy:  Fix redist static and connected routes to ospf.
Fix has been implemented and tested in two cases:
 1. redistributing static routes to ospf4 and ospf6.
 2. redistributing connected routes to ospf4 and ospf6.

In both two cases, the improper behavior was gone. So this is a generic
fix not limited to specific protocol. Based on the current policy code
architecture, I think this is a reasonably fix to the root cause.

From:  Li Zhao

Signed-off-by: Ben Greear <greearb@candelatech.com>