fea:  Add work-around for older kernels that cannot pull single interfaces.
One of the optimizations I added in xorp.ct was to only pull information
for devices we are configured to use, instead of asking netlink for the
entire tree.  This saves a lot of work when you have 1000 interfaces and
xorp only configured for a few of them.

However, older kernels (2.6.18, when compile without
CONFIG_NET_WIRELESS_RTNETLINK, for example) cannot get netlink info
for a single device.

So, this patch adds a work-around that allows xorp to fall back to
grab all interfaces for kernels that only support this method.

Thanks to saurabh.pandya@elitecore.com for reporting this bug and
giving me enough info to track it down.

Signed-off-by: Ben Greear <greearb@candelatech.com>