IPMR:  Add support for new multiple mcast table API.
This new API is the proposed solution we are trying to get
into the upstream Linux kernel.  Leaving in the old API
for now to ease migration for anyone using the old API.

All of this is dynamically discovered at run-time, so
it should be backwards/fowards compatible with whatever
kernel support exists.

NOTE:  Getting this to work properly currently requires patches to 'ip'
from the iproute2 package, as well as the Linux kernel changes.

Signed-off-by: Ben Greear <greearb@candelatech.com>