fea:  Handle funky input from netlink without crashing.
We found a situation where we were evidently getting a bad
subnet prefix (181) reported by netlink API.  This was causing
an un-caught exception that crashed FEA.

I think I found the right place to catch this exception, but
now I cannot reproduce the failure.  Leaving in a bit of debug
print statements in case this crops up again.

Signed-off-by: Ben Greear <greearb@candelatech.com>