libxorp/heap:  Fix long-standing memory corruption.
We used to leak Heap objects on purpose because destroying
them caused crashes.  It seems the crash was either because
of un-initialized memory or (more likely I think), using
memset() on this in the constructor.  Either way, this
patch appears to fix the problem and we run clean under
valgrind (as far as Heaps are concerned, at least).

Signed-off-by: Ben Greear <greearb@candelatech.com>