fea/firewall:  Enable firewall support on linux.
Was missing some logic in the conversion to scons it seems.

Also, you have to hack on
/usr/include/linux/netfilter_ipv6/ip6_tables.h, line 222 or so
and
/usr/include/linux/netfilter_ipv4/ip_tables.h, line 282 or so

to make the methods look something like this:

/* Helper functions */
static __inline__ struct ipt_entry_target *
ipt_get_target(struct ipt_entry *e)
{
        /* BEN:  Was void* */
        return (struct ipt_entry_target *)e + e->target_offset;
}

I'll try to get a fix to the linux kernel headers pushed upstream,
but that could take a long time.

Signed-off-by: Ben Greear <greearb@candelatech.com>