ospf:  Remove assertion when FINDER is lost.
We saw this problem when system was under extreme memory
pressure and CPU load.  Seems finder timed out and was deleted,
and then the assert happened due to NO_FINDER in ospf's send_cb()
method.

Change the assert to just log the error and continue, it doesn't
look much worse than the other errors that are handled similarly.

Signed-off-by: Ben Greear <greearb@candelatech.com>