pim/igmp:  Fix config-file reload issues.
There are funny races that cause pim and igmp to try
to enable (and otherwise configure) interfaces before they
are added when you load a config file with new interfaces
in it.

This patch works around these issues and it now seems to work
properly.  Removed useless comments and a bit of other code cleanup
as well.

Signed-off-by: Ben Greear <greearb@candelatech.com>