vlans: Treat vlans as primary interfaces.
The vlan information is stored on the interface, not vif:

Syntax for vlan 5 on eth0:

interface eth0.5 {
  parent-ifname: "eth0"
  iface-type: "VLAN"
  vid: "5"
  vif eth0.5 {
  }
}

In addition:  Consolidate all the vlan-helper classes into one, and
use ifdefs for various platform code (bsd and linux).  This saves
a lot of code duplication, at the cost of a bit more ifdef complexity
in the remaining files.

Signed-off-by: Ben Greear <greearb@candelatech.com>