bsd:  Fix VLAN creation and deletion.
There was a bug in deletion logic that I introduced
recently:  It would recurse until it ran out of stack
and then crash.

There were more bugs with creating interfaces:  The ioctl call
to configure the VID was returning EBUSY.  Changing to
use system("ifconfig ed0.5 create") fixes this problem,
so ignoring root cause of that ioctl failure.

Fix a third bug that caused xorp to assert and crash when
a vlan device (or any other device) is removed.

Signed-off-by: Ben Greear <greearb@candelatech.com>