ospf:  Deal better with reloading ospf config files.
Problem is this:  When loading a config file that has a new
interface, the route redistribution messages can come before
the new interface info is pushed.  That could lead to trying
to announce routes when the peer has a zero prefix.

This attempts to fix this by basically removing the assert.
I am hoping that the route redistribution will happen properly
when the peer comes up, but I'm not certain that is the case.
So, maybe more fixes coming..but at least we don't crash now.

Core was generated by `xorp_ospfv2'.
Program terminated with signal 6, Aborted.
(gdb) bt
    at libxorp/xlog.c:480
    a3=0xffffffffffffffff <Address 0xffffffffffffffff out of bounds>, a4=140733526888592) at ./libxorp/callback_nodebug.hh:8966

Signed-off-by: Ben Greear <greearb@candelatech.com>