pim:  Use pointers instead of references for members.
pim has a nasty bug that causes pure virtual functions to
be called during some destructors.  This is probably due to
the tangle of class references held in various structs.

I believe the way to clean this up is to use pointers and
then carefully null them out on teardown and add appropriate
checks for null in the destruct sequences.

This patch should make no functional change..just change to
using pointers.

Signed-off-by: Ben Greear <greearb@candelatech.com>