pim:  Fix recently added bugs.
There were two bugs in the code I recently added to pim.  First, the
order of variables in PimMre is vital:  You have to initialize pim_mrt
before you can init some of the others (which use _pim_mrt).

The DELETE_BUFFER call needs a check for null.

Also, re-ordered the class so that all the data members were
private so that it is easier to see data v/s methods in
the header file.

Signed-off-by: Ben Greear <greearb@candelatech.com>