fea:  Fix GRE tunnel IP addresses on Linux.
The old code used the 'link-level' IP address, ie the IP
that GRE is using underneath, as one of the GRE interface's
IP addresses.  I see no reason to do this, and it appears to
mess up multicast over GRE tunnels.

Since there were two addresses, it was a bit of a race as to which
one won, which might be why sometimes multicast over GRE
somewhat worked.

Reported-by: J G Miller <miller@yoyo.ORG>
Signed-off-by: Ben Greear <greearb@candelatech.com>