build-info:  Add xorp version string to xorpsh.
This also makes the version available to all code,
so remove the #define hack in master_conf_tree.cc

This means one less place we have to change the xorp
version.

Also, add run-time info (uname -a, cat /etc/issue if it exists)
to the xorp logs to aid trouble-shooting bug reports.

Signed-off-by: Ben Greear <greearb@candelatech.com>