ospf:  Fix bad lsa retransmit timer bug.
A user reported very slow OSPF convergence times when
xorp was connected to a larger OSPF network.  The root
cause is that the LSA retransmit timer could get in a state
where it rarely if ever fired due to it being constantly
reset whenever an LSA arrived.

It would eventually go to Full when all of the LSAs were
(re)flooded to it, but this could take many minutes.

The fix is to not restart the timer if it is already running.

This bug appears to be quite old.

Signed-off-by: Ben Greear <greearb@candelatech.com>