fea/vlans:  Don't delete vlans unless xorp created them.
Xorp can create 802.1q VLANs, and if so, it will set a flag
internally marking the vlan as being created by xorp.

When an 802.1q vlan interface is removed from xorp the
interface will be deleted from the OS if and only if the
VLAN was flagged as created by xorp.

If this restriction proves to be a problem for anyone, I can
add an environ-variable that will allow forcing delete of all
VLANs and/or perhaps disable VLAN deletion entirely.

Signed-off-by: Ben Greear <greearb@candelatech.com>