xorp: Introduced new flag enable_builddirrun for compiling Xorp
Introduced enable_builddirrun flag, so /lib/xorp/{bin, sbin}
directories would be built even if rtld_origin is set to false
or if shared is set to false.
In this folder module and tool binaries are placed so Xorp
could be run from BUILD_DIR
By default this flag is set to false.
Changed BUILD_NOTES according to this.

Signed-off-by: Igor Maravic <igorm@etf.rs>