xorp: rtrmgr: Fix using of return varibles in xrl
Without this patch, when we read and write from/to same xrl variable,
value that should be committed to variable is overwritten with fake value.

For example:
...
parent @: uint {

	%create: xrl "$(prot.targetname)/prot/0.1/create?node:u32=$(@)";
	%activate: xrl "$(prot.targetname)/prot/0.1/activate?node:u32=$(@)->child_ret:u32=$(@.child)";
	%update: xrl "$(prot.targetname)/prot/0.1/update?node:u32=$(@)->child_ret:u32=$(@.child)";

	child {
		%set: xrl "$(prot.targetname)/prot/0.1/set?node:u32=$(parent.@)&child:u32=$(@)";
	}
}
...

In this case in function set we would get bogus value for variable child, because we are writing to it
from update functions.

This patch fixes that.

Case when we are calling %create + %set + %activate is working without this patch.

Signed-off-by: Igor Maravic <igorm@etf.rs>