xorp: rtrmgr: Include new command %unique-in
With %unique-in command we can set the scope in which our variables are unique.
It takes only on parameter - one of variable parents (or some older ancestor).

Before commit can start it will be checked if the variable is unique in given scope.
If it isn't, error will be returned.

It's only valid for leaf values.

Example template file:

grand-parent @: u32 {
    parent @: u32 {
        child: u32;
    }
}

grand-parent @: u32 {
   %create ...
   %update ...

   parent @: u32 {
       %create ...
       %update ...

       child {
           %unique-in: $(grand-parent.@);

           %set ...
       }

   }

}

This template file means that all child nodes in grand-parent's scope should have unique values.

With this template file this is allowed configuration:

grand-parent 2 {
   parent 2 {
      child: 3
   }
   parent 3 {
      child: 4
   }
}

grand-parent 3 {
   parent 2 {
       child: 4
   }
   parent 4 {
       child: 5
   }
}

This is not allowed configuration:

grand-parent 2 {
   parent 2 {
      child: 4
   }
   parent 3 {
      child: 4
   }
}

grand-parent 3 {
   parent 2 {
       child: 5
   }
}

Signed-off-by: Igor Maravic <igorm@etf.rs>