xorp: ospf: Make checksumming more efficient
Instead of getting checksum out of OSPF header, zeroing checksum location, calculating checksum and
comparing if computed and received checksum are equal, just checksum received packet and see if result
is equal to 0.
Based on: http://en.wikipedia.org/wiki/Header_checksum

In case the received checksum is invalid, calculate it and print it out in exception message.

Signed-off-by: Igor Maravic <igorm@etf.rs>