xorp: rib: Enable compiling of rib tests
Tests are compiled to the {$builddir}/rib/tests
Files commands, test_rib_xrls.sh and
test_rib_direct.sh are copied to that directory also.

They are going to be compiled only if test building is enabled via
scons command line.

Added new enviroment 'builddir' to the SCons. It represents
string representation of build directory.
I needed it so the files could be copied to apropriate dir.

Signed-off-by: Igor Maravic <igorm@etf.rs>