xorp: libxorp: Fix find_subtree function of TrieNode
Problem was that it was going unconditionaly right, if the
left branch didn't contain network we were looking for.
That is wrong, because it is posible that left branch is contained
in the network that we're looking for. Because of that it's
possible that it's our candidate.
Added checks for the left and right branches.
Added test that really test if we found correct subtree.
Passed trie test.