xorp: rib: Don't propagate route copy from PolicyConnectedTable
All routes are saved in trie, as pointers to non const objects.
That way they can be modified.
All routes will be cached in subsequent tables, and because of that,
we must pass pointers to objects that won't be destroyed.
This table is responsible of freeing the memory of route objects.
Passed all rib tests.