xorp: rib: Key for saving resolved route entries isn't pointer to IPRouteEntry any more, it's IPNet
When the connected route is deleted, it's first deleted in PolicyConnectedTable and then after that
it's propagated (as different pointer) to other tables.
If the external route is resolved to connected nexthop, and the connected nexthop is deleted,
problem occures because we can't delete the resolved nexthop, because we're searching with different pointer.

Passed all BGP and RIB tests.

Signed-off-by: Igor Maravic <igorm@etf.rs>