xorp: rip: Force RIP to do EXPORT filtering when the route comes
Also, don't check if metric is larger then RIP_INFINITY, before the route passes through the policy filters.
EXPORT and IMPORT filter, could posibly change metric and make it smaller then RIP_INFINITY.

After the route passes through the EXPORT filter, or if it doesn't pass IMPORT or EXPORT_SOURCEMATCH filters,
check the route cost. If it's larger then RIP_INFINITY set it to ROUTE_INFINITY.

Don't run EXPORT filter when the outgoing packet is sent, because route is pushed through the EXPORT filter when it comes to RIP.

Signed-off-by: Igor Maravic <igorm@etf.rs>