rib:  Fix ref-counting in RouteEntry class.
RouteEntry does ref-counting on the vif object, but it did not
implement copy constructors or operator=.  This means ref-counting
didn't work right.  Attempt to fix this by properly implementing
the copy constructors and operator= methods in RouteEntry and
child classes.

RIB is still a scary mess of pointers, but maybe this helps a bit.

Signed-off-by: Ben Greear <greearb@candelatech.com>